chore: move Queue and interfaces to infrastructure
