+++ src/ATech.Ring.DotNet.Cli/Infrastructure/IReceiver.cs
@@ -1,8 +1,9 @@
-﻿namespace ATech.Ring.Protocol.v2;
-
-using System;
+﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using ATech.Ring.Protocol.v2;
+
+namespace ATech.Ring.DotNet.Cli.Infrastructure;
 
 public delegate Task OnDequeue(Message message);
