+++ src/ATech.Ring.DotNet.Cli/Infrastructure/ISender.cs
@@ -1,7 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using ATech.Ring.Protocol.v2;
 
-namespace ATech.Ring.Protocol.v2;
+namespace ATech.Ring.DotNet.Cli.Infrastructure;
 
 public interface ISender
 {
