+++ src/ATech.Ring.DotNet.Cli/Program.cs
@@ -14,7 +14,6 @@ using ATech.Ring.DotNet.Cli.Infrastructure;
 using ATech.Ring.DotNet.Cli.Infrastructure.Cli;
 using ATech.Ring.DotNet.Cli.Logging;
 using ATech.Ring.DotNet.Cli.Workspace;
-using ATech.Ring.Protocol.v2;
 using k8s;
 using LightInject;
 using LightInject.Microsoft.AspNetCore.Hosting;
@@ -77,9 +76,9 @@ try
     builder.Services.AddSingleton<IWorkspaceLauncher, WorkspaceLauncher>();
     builder.Services.AddSingleton<IWorkspaceInitHook, WorkspaceInitHook>();
     builder.Services.AddSingleton<ICloneMaker, CloneMaker>();
-    builder.Services.AddSingleton<ATech.Ring.Protocol.v2.Queue>();
-    builder.Services.AddSingleton<ISender>(f => f.GetRequiredService<ATech.Ring.Protocol.v2.Queue>());
-    builder.Services.AddSingleton<IReceiver>(f => f.GetRequiredService<ATech.Ring.Protocol.v2.Queue>());
+    builder.Services.AddSingleton<Queue>();
+    builder.Services.AddSingleton<ISender>(f => f.GetRequiredService<Queue>());
+    builder.Services.AddSingleton<IReceiver>(f => f.GetRequiredService<Queue>());
     builder.Services.AddSingleton(f =>
     {
         var configuredPath = f.GetRequiredService<IOptions<RingConfiguration>>().Value.Kubernetes.ConfigPath;
