+++ src/ATech.Ring.DotNet.Cli/Runnables/Dotnet/AspNetCoreRunnable.cs
@@ -2,8 +2,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using ATech.Ring.Configuration.Runnables;
+using ATech.Ring.DotNet.Cli.Infrastructure;
 using ATech.Ring.DotNet.Cli.Tools;
-using ATech.Ring.Protocol.v2;
 using Microsoft.Extensions.Logging;
 using static ATech.Ring.DotNet.Cli.Dtos.DetailsKeys;
