+++ src/ATech.Ring.DotNet.Cli/Runnables/Kustomize/KustomizeRunnable.cs
@@ -8,7 +8,6 @@ using ATech.Ring.DotNet.Cli.Abstractions;
 using ATech.Ring.DotNet.Cli.Dtos;
 using ATech.Ring.DotNet.Cli.Infrastructure;
 using ATech.Ring.DotNet.Cli.Tools;
-using ATech.Ring.Protocol.v2;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using static ATech.Ring.DotNet.Cli.Tools.ToolExtensions;
