+++ src/ATech.Ring.DotNet.Cli/Runnables/Windows/NetExe/NetExeRunnable.cs
@@ -2,8 +2,8 @@
 using System.Threading.Tasks;
 using ATech.Ring.DotNet.Cli.CsProj;
 using ATech.Ring.DotNet.Cli.Dtos;
+using ATech.Ring.DotNet.Cli.Infrastructure;
 using ATech.Ring.DotNet.Cli.Tools;
-using ATech.Ring.Protocol.v2;
 using Microsoft.Extensions.Logging;
 using NetExeConfig = ATech.Ring.Configuration.Runnables.NetExe;
