+++ src/ATech.Ring.DotNet.Cli/Tools/GitClone.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Data;
 using System.IO;
 using System.Text.RegularExpressions;
 using System.Threading;
